refactor: Github API를 호출하는 로직을 트랜잭션 범위 밖으로 분리
